src/requirements.txt
amqp==2.2.2
anyjson==0.3.3
appnope==0.1.0
Babel==2.5.3
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.3
bleach==2.1.3
celery==4.0.2
certifi==2018.1.18
chardet==3.0.4
configparser==3.5.0
coreapi==2.3.3
coreschema==0.0.4
decorator==4.2.1
Django==1.11.11
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-extensions==2.0.6
django-filter==1.1.0
django-picklefield==1.0.0
django-rest-swagger==2.1.2
djangorestframework==3.7.7
djangorestframework-csv==2.1.0
entrypoints==0.2.3
enum34==1.1.6
flower==0.9.2
functools32==3.2.3.post2
futures==3.2.0
html5lib==1.0.1
httpie==0.9.9
idna==2.6
ipykernel==4.8.2
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.1.2
itypes==1.1.0
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kombu==4.1.0
Markdown==2.6.11
MarkupSafe==1.0
mistune==0.8.3
mysqlclient==1.3.12
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.1
numpy==1.14.2
openapi-codec==1.3.2
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pathlib2==2.3.0
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.7.0
pytz==2018.3
pyzmq==17.0.0
qtconsole==4.3.1
red==0.2.1
redis==2.10.6
requests==2.18.4
scandir==1.7
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.13.2
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.2.6
terminado==0.8.1
testpath==0.3.1
tornado==5.0.1
traitlets==4.3.2
typing==3.6.4
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.4
xlrd==1.1.0
